+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 - Add new groupsList runtime function.
 - Add runtime leaderboardList and leaderboardsGetId functions.
 - Add leaderboard/tournament prev_reset field.
+- Add custom metrics runtime functions for counters, gauges, and timers.
 
 ### Changed
 - Include ticket in party matchmaker add operation responses.